Applied fixes for vector inputs
- attempt to support both 1-dimensional and 2-dimensional vectors as input

- added special handling for SymPy, which doesn't support 1-dimensional vectors

+++ python/tests/test_softmax.py
@@ -184,17 +184,17 @@ class TestSoftmaxValues(TestCase):
         X = np.array([
             [1, 2, 3],
             [7, 3, 4]
-        ], dtype=float)
+        ], dtype=np.float32)
 
         xc = np.array([
             [9],
             [3],
             [12],
-        ], dtype=float)
+        ], dtype=np.float32)
 
         xr = np.array([
             [11, 2, 3]
-        ], dtype=float)
+        ], dtype=np.float32)
 
         return X, xc, xr
